// packages/shared/src/types/skill.ts
export enum SkillName {
  MELEE = "melee",
  RANGED = "ranged",
  DEFENSE = "defense",
  AGILITY = "agility",
  FISHING = "fishing",
  MINING = "mining",
  WOODCUTTING = "woodcutting",
  FORAGING = "foraging",
  COOKING = "cooking",
  SMITHING = "smithing",
  ALCHEMY = "alchemy",
}

export const ALL_SKILLS = Object.values(SkillName);
export const MAX_SKILL_LEVEL = 99;